refactor: songbook types
