+++ constants.ts
@@ -1,1 +1,1 @@
-export const SCORES_COLLECTION = "scores";
+export const SCORES_COLLECTION = "songs";
